Make the game playable in console
