+++ index.js
@@ -1,3 +1,2 @@
-// This is where your project starts.
-
-console.log('Your project is running...'); 
+const {start} = require('./src/game')
+start();
