+++ src/game.js
@@ -1,4 +1,6 @@
+const { createCard } = require('./card');
 const data = require('./data');
+const { createDeck, createRound, countCards } = require('./round');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
 
@@ -11,4 +13,21 @@ function printQuestion(round) {
   util.main(round);
 }
 
-module.exports = { printMessage, printQuestion };
+const makeCards = cardsData => {
+  return cardsData.map(cardData => createCard(cardData.id, cardData.question, cardData.answers, cardData.correctAnswer));
+}
+
+const start = () => {
+  const cards = makeCards(prototypeQuestions);
+  const deck = createDeck(cards);
+  const roundZero = createRound(deck);
+  printMessage(deck);
+  printQuestion(roundZero);
+}
+
+module.exports = { 
+  printMessage, 
+  printQuestion,
+  makeCards,
+  start 
+};
